requirements.txt
asgiref==3.3.1
Django==3.1.6
django-nose==1.4.7
djangorestframework==3.12.2
nose==1.3.7
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
